feat(3-function_like_macro.h): Add ABS function-like macro for computing absolute value
Implement the ABS function-like macro in the 3-function_like_macro.h header file. This macro calculates the absolute value of a number using a ternary operator. It checks if the number is less than 0 and returns the negative value if true, otherwise, it returns the number itself. The ABS macro provides a convenient way to compute absolute values in the codebase.