feat(free_listint2): Implement function to free a listint_t list and set head to NULL
This commit adds the implementation of the  free_listint2  function in the  5-free_listint2.c  file. The function takes a double pointer to the head of a linked list of type  listint_t  as a parameter. It frees all the nodes in the list and sets the head pointer to NULL.

The  5-main.c  file is also included in this commit, which provides a  main  function to test the functionality of the  free_listint2  function.

Additionally, the  lists.h  header file is updated to include the prototype of the  free_listint2  function.

The code has been tested and is functioning as expected, successfully freeing the memory allocated for the linked list and setting the head pointer to NULL.

This commit strictly adheres to the provided requirements and guidelines for this task, including the use of block comments, adherence to the Betty style, and the inclusion of the necessary header file  lists.h  that contains the definition of the  listint_t  structure and the updated function prototype.