feat(clear_bit): implement function to clear a bit at a given index
This commit adds the implementation of the clear_bit function, which sets the value of a bit to 0 at a specified index. The function takes a pointer to an unsigned long int and an index as parameters. It checks if the index is within the range of the number's bits, and if so, it clears the bit at that index by performing a bitwise AND operation with the complement of a bit mask. The commit also includes the necessary header file updates.