feat(hash function): Implement hash_djb2
Implement the hash_djb2 function in hash_tables.h to calculate the hash value of a given string using the djb2 algorithm. The function efficiently calculates the hash by iterating over each character in the string and updating the hash value. The calculated hash is returned as an unsigned long int.