chore(hash_table): Add key_index function
Add key_index function to calculate the index at which a key/value pair should be stored in an array of a hash table. The function uses the djb2 algorithm and takes the key and size of the array as parameters. It returns the index of the key in the array.